+++ digest/src/lib.rs
@@ -205,6 +205,27 @@ pub trait CustomizedInit: Sized {
     fn new_customized(customization: &[u8]) -> Self;
 }
 
+/// Trait for hash functions with customization string for domain separation which place
+/// restrictions on customization strings.
+pub trait TryCustomizedInit: Sized {
+    /// Error returned for invalid customization strings.
+    type Error;
+
+    /// Create new hasher instance with the given customization string.
+    ///
+    /// # Errors
+    /// If the provided customization string is not valid for the hash function.
+    fn try_new_customized(customization: &[u8]) -> Result<Self, Self::Error>;
+}
+
+impl<T: CustomizedInit> TryCustomizedInit for T {
+    type Error = core::convert::Infallible;
+
+    fn try_new_customized(customization: &[u8]) -> Result<Self, Self::Error> {
+        Ok(Self::new_customized(customization))
+    }
+}
+
 /// Types with a certain collision resistance.
 pub trait CollisionResistance {
     /// Collision resistance in bytes.